u32 for everything

+++ assembler/src/generator/st.rs
@@ -121,27 +121,27 @@ fn generate_st_reg_ireg_offset(
     modifiers: &Spanned<Vec<Spanned<Modifier>>>,
     register1: &Register,
     register2: &Register,
-    offset: i32,
+    offset: u32,
 ) -> Result<u32, Error> {
     let mut opcode: u32 = 0;
     opcode |= generate_modifiers_non_alu(modifiers)?;
     opcode |= Mnemonic::STR.generate();
     opcode |= register1.generate() << 16;
     opcode |= register2.generate() << 12;
-    opcode |= offset as u32 & 0xfff;
+    opcode |= offset & 0xfff;
     return Ok(opcode);
 }
 
 fn generate_st_reg_inum(
     modifiers: &Spanned<Vec<Spanned<Modifier>>>,
     register: &Register,
-    number: &Spanned<i32>,
+    number: &Spanned<u32>,
 ) -> Result<u32, Error> {
     assert_bit_length(&number, 16)?;
     let mut opcode: u32 = 0;
     opcode |= generate_modifiers_non_alu(modifiers)?;
     opcode |= Mnemonic::ST.generate();
     opcode |= register.generate() << 16;
-    opcode |= number.val as u32 & 0xffff;
+    opcode |= number.val & 0xffff;
     return Ok(opcode);
 }
